Add animation for when user clicks the favorite button
When the user clicks the favorite button, the favorite star will scale
up or down depending on its new favorite status. A dark gray favorite
star will remain in either case. When the star is scaled up, the gray
star will show as a outline behind the new yellow star. If the star is
scaled down, the gray star will remain to denote its non-favorite
status.

The favorite status is toggled after the animation has played from
within RecipeItemAnimator. This prevents the star from showing suddenly
when onBindViewHolder is called from within AdapterRecipe.

Fixed bug where notifyItemChanged() in the RecipeAdapter would cause
additional views to be added in the background.
 - Resolves bug from last commit (not tracked)
 - In tablet-portrait mode, app will still occasionally crash when
   unable to find the fragment container for some reason. Not easily
   reproducible (see #7)