Add animation for when recipes are initially loaded
Added new animation for RecipeItemAnimator for when recipes are first
loaded into the RecyclerView. Items will enter from bottom and move to
their respective locations using a translation animation. AdapterRecipe
detects when a new Cursor is to replace a null mCursor to infer that the
list was previously empty. It then utilizes notifyItemRangeInserted to
force the ItemAnimator to utilize the animation from animateAdd.

Also fixed bug where the "Direction" section header would be showing on
the device even though all views were supposed to be invisible while
importing the recipe. The TextView was added to a larger transition
group who's animation was correctly handled during the window
transition.