Improve performance using master-detail flow view
Discovered that utilizing a LinearLayout and layout weights caused the
system to attempt to draw all ViewHolders in AdapterRecipe
simultaneously every time there was a change including:
 - notifyDataSetChanged()
 - Scrolling
 - Selecting a recipe
 - Closing a recipe

Modified the layout to utilize RelativeLayout and an invisible aligner
to avoid using layout weights. Also changed from setting the detail
view to View.GONE to setting its width to 0 work with the aligner.

The RecyclerView now smooth-scrolls to the selected recipe in master-
detail flow view to enhance UI.