Activate search feature
Worked out most issues with searching each of the sources for the
user-input search term. AllRecipes, Epicurious, and Food.com are all
queried simultaneously for the search term and the results are shown in
AdapterRecipe, but not added to database to not flood the main page with
the search results.

TODO:
 - Add ability to click on recipe to view details
 - Fix bug where search term containing more than one word will only
   return results from Food.com (see #10)