Modify onClickHandler in AdapterRecipe to use recipe URL
The function originally takes a long recipeId parameter, but in almost
every case, it is converted to a recipe URL before being passed to the
Activity to open ActivityRecipeDetails. Modified it to remove the extra
step and allow compatibility in ActivitySearch as recipes do not have a
recipeId yet to pass to the onClickHandler.