Alter ActivityRecipeDetails layout
Now using CollapsingToolbarLayout that includes the recipe image. The
image view uses parallax scrolling to hide the image view as the
NestedScrollView holding the Fragment is scrolled up. The toolbar itself
will change from transparent to colored as the user scrolls down to
remain visible.

Also added Snackbar to inform the user of lack of network connectivity
when there are already recipes in the database in ActivityRecipeList.

Added Snackbar to inform user of lack of network connectivity when
attempting to view recipe details that have not been imported yet.