Implement NavigationDrawer on all main Activities for UI consistency
Altered the layout of:
 - Favorites
 - My Recipes
 - Recipe Books
to include the NavigationDrawer from ActivityRecipeList so that
navigation within the app is consistent and easy. Back button will
always take the user back to ActivityRecipeList from these auxiliary
Activities. The current Activity is always highlighted in the navigation
menu to indicate to the user. Clicking on the current Activity will do
nothing.

Also added scrolling behavior to the toolbar of these auxiliary
Activities.