Fix ActivityCreateRecipe bugs
Reworked ActivityCreateRecipe so that deleting recipes will maintain
correct recipe ID.

Added icon to indicate that an image needs to be added.

Soft-keyboard now shows when user clicks the button to add a new
ingredient or direction.

Pressing the IME next action automatically shifts focus to the
description EditText.

Added ability to change the author name of custom recipes if it hasn't
already been set. A Dialog pops up to request the user input their
name if it detects that no author name has yet been set.

When user selects option to save the recipe to database, it
automatically launches a new ActivityRecipeDetails showing the recipe's
details.