Create ActivityShoppingList for displaying shopping list
Copied the layout and function of ActivityFavorites. Added the new
Activity as a menu option in the NavigationDrawer of all relevant
Activities. The Activity utilizes a CursorLoader filtering the
LinkIngredient table for all ingredients that should be in the shopping
list.

Modified AdapterIngredient to allow for displaying the title of a recipe
before the list of ingredients needed for it in ActivityShoppingList.