Add checks when trying to add a recipe to a recipe book/chapter
In ActivityRecipeDetails, when the user selects to add a recipe to the
recipe book, there was no way to tell whether a chapter or recipe book
already contained it. A dialog was added to warn a user that a recipe
book already contains a chapter, asking if they would like to continue.
Then if the user attempts to add a recipe to a chapter that already
contains the recipe, a Toast is shown to inform the user to select a
different chapter.