Add AuthorityMatcher
Testing with API 16 phones showed an error in using the URIMatcher to
identify URIs due to a weird rule where it would only match the first
URI with a partial match regardless of whether following URIs would be a
complete match. Replaced it with a custom-made AuthorityMatcher that
only attempts to match the authority of the given URI as that is all
that is required for the purpose of this app.