Fix bug causing ActivityShoppingList to crash
Previously, if more than one recipe's worth of ingredients were added to
ActivityShoppingList, the app would crash immediately. This was caused
by a modifier that should have been applied after a loop that was placed
within the loop.

Note* I'm not sure when this change occurred because previous builds had
the function working properly with multiple recipes.