Add ability to backup and restore database
Added an option in SettingsActivity to allow the user to create a local
backup or create a backup that is uploaded to Google Drive. Also
includes options to restore from either of those locations. This will
allow the user to backup their favorites, custom recipes, and recipe
books across multiple devices and guard against catastrophic failure.

Will be implementing a sync feature in later builds that will allow the
user to have their database synced across multiple devices.