Remove hard-coded custom-prefix
Previously, custom recipes were hard-coded with the asterisk prefix.
Replaced it with a reference to a resource so that it can be easily
changed in the future if required.

Also created a method for performing tasks that should be run once for
one-time changes. I.e. changing the prefix of all custom recipes from
an asterisk to the new custom-prefix.